Rename s4_slit_gaps module to slits
