+++ src/dodal/beamlines/p38.py
@@ -6,7 +6,7 @@ from dodal.beamlines.beamline_utils import (
 from dodal.beamlines.beamline_utils import set_beamline as set_utils_beamline
 from dodal.common.visit import StaticVisitDirectoryProvider
 from dodal.devices.areadetector import AdAravisDetector
-from dodal.devices.s4_slit_gaps import Slits
+from dodal.devices.slits import Slits
 from dodal.devices.tetramm import TetrammDetector
 from dodal.log import set_beamline as set_log_beamline
 from dodal.utils import get_beamline_name
